migration/requirements.txt
python-hcl2>=4.3.2
click>=8.1.7
pathlib>=1.0.1
Jinja2>=3.1.3
